- Fixed bug in policy edit/creation, if node limitations were added the other limitation values would be removed,   this no longer happpens. - Fixed bug in listing of subtree when the attribute filter and attribute sort were used and the attribute filter   was an empty array. - Fixed security bug in content module for several views, they now have a separate function access instead   of reusing the 'edit' access. - Fixed security bug in policy checking for content objects, it was possible to get create access   anywhere on the site all depeneding on the role setup. Now creation checking is properly implemented. - Fixed security bug in which it was possible to add forbidden locations to an object you could edit. - Fixed security bug in which any object could be created using the related object functionality for   an existing object edit.
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/stable/3.2@4591 a01eee8c-daba-0310-acae-fa49f3023285

+++ kernel/content/removeeditversion.php
@@ -49,6 +49,8 @@ if ( $http->hasPostVariable( "ConfirmButton" ) )
     $object =& eZContentObject::fetch( $objectID );
     if ( $object === null )
         return $Module->handleError( EZ_ERROR_KERNEL_NOT_AVAILABLE, 'kernel' );
+    if ( !$object->attribute( 'can_edit' ) )
+        return $Module->handleError( EZ_ERROR_KERNEL_ACCESS_DENIED, 'kernel' );
     $db =& eZDB::instance();
     $db->query( "DELETE FROM ezcontentobject_link
 		         WHERE from_contentobject_id=$objectID AND from_contentobject_version=$version" );
